DOCS Updating information about BlogMigrationTask
This no longer automatically runs during dev/build. Migration should be conscious decision to run when a developer decides to update and avoids the risk of accidentally migrating on a live site when perhaps this was not the intention. I mention the move from 1.0 to 2.0 as this is a safe migration that some version old versions. Likely the upgrade path would be best if (1) Upgrade you really old blog up to 1.0 (2) Update your codebase to 2.0 code (3) Run the migration task.